+++ CHANGES.rst
@@ -1,6 +1,9 @@
 1.5 (unreleased)
 ================
 
+- Added the ``source_value`` and ``alt_identifier`` information to the verbose
+  output of ``describe()`` in ``regtap``. [#492]
+
 - Added convenience method DALResults.to_qtable() that returns an
   astropy.table.QTable object. [#384]